# Conflicts:
#	src/main/java/com/StudyMate/StudyMate/controller/FlashCardController.java
#	src/main/java/com/StudyMate/StudyMate/model/Background.java
#	src/main/java/com/StudyMate/StudyMate/repository/UserRepository.java
#	src/main/java/com/StudyMate/StudyMate/serviceImpl/FavouriteBackgroundImpl.java
#	src/main/java/com/StudyMate/StudyMate/serviceImpl/UserServiceImpl.java

+++ src/main/java/com/StudyMate/StudyMate/service/UserService.java
@@ -5,9 +5,9 @@ import com.StudyMate.StudyMate.model.User;
 import java.util.List;
 
 public interface UserService {
-    List<User> getBackgroundByUserId(long userId);
+    List<User> getBackgroundByUserId(int userId);
 
-    void delete(long userId);
+    void delete(int userId);
 
     void save(User user);
 }
